feat: add prune_notification_store endpoint to QStashClient (#243)
* feat: add prune_notification_store endpoint to QStashClient

This commit introduces a new asynchronous method, `prune_notification_store`, to the `QStashClient` for managing notification storage. Additionally, it updates the routing in `mod.rs` to expose this functionality via a new API endpoint.

* feat: spawn prune_notification_store task in QStashClient initialization

This commit modifies the QStashClient constructor to spawn an asynchronous task that invokes the `prune_notification_store` method upon initialization. This ensures that the notification store is managed concurrently without blocking the client setup. Additionally, it removes commented-out code from the notification store job file for clarity.

* feat: add Upstash-Method header to QStashClient requests

This commit adds the "Upstash-Method" header with a value of "POST" to the HTTP requests made by the QStashClient. This enhancement ensures that the requests are properly identified as POST requests, improving compatibility with the Upstash service.

* chore: update dependencies and fix notification store pruning error handling

This commit updates several dependencies in the Cargo.lock file, including versions for 'bon', 'darling', 'cc', 'rand', and others. Additionally, it modifies the `prune_notification_store` function in `notification_store_job.rs` to improve error handling by mapping errors to an internal server error response format.

* feat: rename and enhance prune notification store scheduling in QStashClient

This commit renames the `prune_notification_store` method to `register_prune_notification_store_schedule` and adds logic to check for existing schedules before creating a new one. It ensures that duplicate schedules are not created and logs the outcome of the scheduling process.

* refactor: remove unused prune notification store schedule registration from QStashClient

* chore: update dependencies and fix source references in Cargo.lock

---------

Co-authored-by: JoeruCodes <joymedi8143@gmail.com>